requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
python-dotenv==1.0.0
joblib==1.3.2
